+++ Quiz_App/Model/Question.swift
@@ -1,0 +1,13 @@
+//
+//  Question.swift
+//  Quiz_App
+//
+//  Created by utkarsh mishra on 05/07/24.
+//
+
+import Foundation
+
+struct Question{
+    let text: String
+    let answer: String
+}
